Don't hardcode window dimensions in window manager
This can probably be improved so that the fractions aren't as ugly.  I
also fixed a small UI issue where the input window didn't quite reach
the bottom of the screen.  Not sure if I fixed it correctly though.

This should be the last piece of work necessary to facilitate dimension
changes through an options screen.